Autentify/requirements.txt
fastapi
python-jose
passlib
sqlalchemy
mysqlclient
pydantic
uvicorn